cli: move environment check helper to src/utils/environment.ts
Avoids coupling login command to status command module.
